requirements.txt
Authlib==1.4.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.19.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
